Solution to Day 07
I tried writing this in Chapel but I wasn't able to get the recursive
types to work. In another "throwback to grad school" I went with Clojure
instead (which is one of the things I was learning instead of working on
my thesis).

I somehow ened up with ~100 lines of parsing code and then solved the
actual problems in ~20 lines apeice. ':D